hw/vfio/iommufd: Use host PAGE_SIZE
For host kernel that is configured to PAGE_SIZE=64K for example, this
4096 hard-coding won't work.

The iommu_domain must always support the system CPU page size because
there is no reasonable way in Linux for userspace to allocate physical
contiguous pages that would be compatible with a larger minimum page
size.

So simply set to sysconf(_SC_PAGE_SIZE).

Change-Id: I8a36155148b5fee294b9682958200a75fbca05e8
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>