vfio: Rename VFIOContainer into VFIOLegacyContainer
In the prospect to introduce a base object for the VFIOContainer
and derive into the existing legacy container and the iommufd
based contained, let's rename the existing one into
VFIOLegacyContainer. This is just an incremental step to ease
the migration. Soon there won't be any reference to the legacy
container in the common.c code. Only the container.c should
handle the VFIOLegacyContainer object.

Signed-off-by: Eric Auger <eric.auger@redhat.com>