hw/pci: Introduce PCIIOMMUOps::set_host_iova_regions
This new callback will be used to convey usable IOVA regions
from VFIO-PCI devices to vIOMMUS (esp. virtio-iommu). The advantage
is that this callback can be called very early, once the device has
is known to be protected by a vIOMMU, after the get_address_space()
has been called by the parent device. Current solution to convey
IOVA regions relies on IOMMU MR callbacks but this requires an
IOMMU MR to be connected with the VFIO-PCI device which generally
comes with the enablement of the IOMMU MR (vIOMMU protection activated).
The downside is that is comes pretty late and in case of virtio-iommu,
after the probe.

Signed-off-by: Eric Auger <eric.auger@redhat.com>